// turborepo/packages/mantine/v0.3.0/src/nodes/organisms/table/fabricVersions/v0.1.0/v2.0.0/types/Column.d.ts
import { RItem } from "@rk/types"
import { DataTableColumn } from "../lib"

export type ColumnDef200 = DataTableColumn<RItem> & {
    sort?: {
        default: 'asc' | 'desc'
        func?(items: RItem[], direction: 'asc' | 'desc'): RItem[]
    }
    filterFunc?(items: RItem[], value: any): RItem[]
    expander?: boolean
    cellProps?: any
}